# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Plan-It API — content approval platform for agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    post:
      operationId: syncMe
      tags: [users]
      summary: Create or sync current user after Clerk login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSync"
      responses:
        "200":
          description: User synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /share-links:
    post:
      operationId: createShareLink
      tags: [shareLinks]
      summary: Generate a public no-login share link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareLinkInput"
      responses:
        "201":
          description: Share link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLink"
  /preview/{token}:
    get:
      operationId: getPreview
      tags: [shareLinks]
      summary: Get public feed for a share link (no auth required)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public preview feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewFeed"
        "404":
          description: Invalid or expired token
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, role]
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [TEAM_ADMIN, TEAM_EDITOR, CLIENT_APPROVER]
        orgId:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        createdAt:
          type: string
    UserSync:
      type: object
      required: [email]
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [TEAM_ADMIN, TEAM_EDITOR, CLIENT_APPROVER]
    Client:
      type: object
      required: [id, name, orgId, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        orgId:
          type: string
        createdAt:
          type: string
        accountCount:
          type: integer
        pendingCount:
          type: integer
...
    SocialAccount:
      type: object
      required: [id, clientId, platform, handle, createdAt]
      properties:
        id:
          type: string
        clientId:
          type: string
        platform:
          type: string
          enum: [INSTAGRAM, LINKEDIN, TIKTOK, FACEBOOK]
        handle:
          type: string
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        postCount:
          type: integer
        pendingCount:
          type: integer
...
    PostMedia:
      type: object
      required: [id, postId, mediaUrl, mediaType, orderIndex]
      properties:
        id:
          type: string
        postId:
          type: string
        mediaUrl:
          type: string
        mediaType:
          type: string
          enum: [IMAGE, VIDEO]
        orderIndex:
          type: integer
    Comment:
      type: object
      required: [id, postId, userId, body, createdAt]
      properties:
        id:
          type: string
        postId:
          type: string
        userId:
          type: string
        body:
          type: string
        createdAt:
          type: string
        userName:
          type: ["string", "null"]
        userEmail:
          type: string
    Post:
      type: object
      required: [id, socialAccountId, type, status, createdById, createdAt, updatedAt, media, comments]
      properties:
        id:
          type: string
        socialAccountId:
          type: string
        type:
          type: string
          enum: [SINGLE, CAROUSEL, REEL, STORY]
        caption:
          type: ["string", "null"]
        status:
          type: string
          enum: [DRAFT, PENDING_APPROVAL, APPROVED, CHANGES_REQUESTED, SCHEDULED, PUBLISHED]
        scheduledAt:
          type: ["string", "null"]
        createdById:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMedia"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
...
    ShareLink:
      type: object
      required: [id, token, clientId, createdAt, url]
      properties:
        id:
          type: string
        token:
          type: string
        clientId:
          type: string
        socialAccountId:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        url:
          type: string
    ShareLinkInput:
      type: object
      required: [clientId]
      properties:
        clientId:
          type: string
        socialAccountId:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
    PreviewFeed:
      type: object
      required: [client, accounts]
      properties:
        client:
          $ref: "#/components/schemas/Client"
        accounts:
          type: array
          items:
            type: object
            required: [account, posts]
            properties:
              account:
                $ref: "#/components/schemas/SocialAccount"
              posts:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
...